Add files via upload
Added the noise variance estimation and optimized it by simply changing the array to a NumPy array. 
Also made a helper function for the load image function to make it run just a little bit faster. (Runs about 40 milliseconds~ and to test it you have to just enter the file_name yourself). 
Added the timeme and profiler so we can track the memory usage / speed to optimize more efficiently.